+++ pennylane/pauli/__init__.py
@@ -28,4 +28,8 @@ from .utils import (
     pauli_mult,
     pauli_mult_with_phase,
     partition_pauli_group,
+    qwc_rotation,
+    diagonalize_pauli_word,
+    diagonalize_qwc_pauli_words,
+    diagonalize_qwc_groupings,
 )
