+++ pennylane/pauli/grouping/__init__.py
