+++ pennylane/pauli/grouping/group_observables.py
